+++ packages/core/src/shared/index.ts
@@ -3,3 +3,4 @@
  */
 export { createBasicAuthHeader } from "@/shared/utils.ts"
 export { createSyslogMessage } from "@/shared/logger.ts"
+export { fetchAsync } from "@/shared/fetch-async.ts"